requirements.txt
click
polib
pyyaml
requests
six
